requirements.txt
openai>=1.0.0
PyGithub>=2.0.0
gitpython>=3.1.40
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
ruff>=0.1.0
